Update modules.rst
Changes made here rather than in a new PR; added Guillaume per his request, perhaps add others who can test this as well?

Modules on orion are tested (my S4 modules are also tested, so we could use those as well, if that's appropriate, I'm ambivalent).

I'd like to suggest that we drop 'Intel' from the jedi-cmake toolchain names, as that's only appropriate for building on cheyenne. On other hosts, the toolchain is used to set a proper value for mpiexec or mpirun, and is necessary regardless of compiler / MPI module suite. Everywhere else, modules are built with C++14 support in them.